Additional mover script
This script sits in /user/local/sbin along side the original mover and mover.old files.
it adds a second argument for the number of days old.
Find command is changed to only move file types, and return files over ## (Arguement 2) days old.  i.e. Added to the -mtime of the find command.