4 changes from Swarles
First change
I changed all the "while read" lines in age_mover to "while IFS= read -r. This was in relation to a post here where the user had an issue with trailing whitespace not being preserved. My thoughts were that at this point in the script whitespaces should not be due to user error anymore. I haven't thoroughly tested it though, I basically just asked chatgpt what the change would do and it seems to agree that it would only impact the trailing and leading whitespaces, but it has been wrong before.

Second change
This was the issue from ages ago where peoples mover would not run. To do with using the mover.old script.

Third change
This prints a log if "share.cfg" doesn't exist and exits so the user knows what is causing the mover to not run properly.

Fourth change
This checks to see if the "ca.mover.tuning.cfg" file contains the relevant information after it has been parsed in mover.php. If it doesn't contain the "moverDisabled" field, it interprets this as the ca.mover.tuning.cfg either not existing or not containing the correct information, creates a log for the user to check it and then it exits.